+++ CHANGELOG.md
@@ -22,10 +22,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [SIL.Core.Desktop] Serializable class `UpdateSettings` (settings for getting updates)
 - [SIL.Windows.Forms] `CssLinkHref` property to `ShowReleaseNotesDialog` to allow linking to CSS file for
     displaying Markdown output.
+- [SIL.Scripture] IScrVerseRef interface (largely extracted from VerseRef)
 
 ### Changed
 
 - [SIL.WritingSystems] Update langtags.json to the latest
+- [SIL.Scripture] Made VerseRef class implement new IScrVerseRef interface
+- [SIL.Forms.Scripture] Changed VerseControl to use IScrVerseRef and not depend directly on ScrVers
 
 ### Fixed